Hotfix: Cleanup descriptions for connection and redis spy tests
This cleans up the assert context descriptions for the connection
and redis spy tests. The redis spy tests were inheriting the
connection spy tests description but they no longer should because
redis spy is not under connection spy. This fixes the contexts,
breaking out a simple unit tests that holds the common
configuration and having the connection and redis spy tests inherit
from it. This should have been part of the last changes that made
the connection and redis spy easier to extend.

/cc @kellyredding