Hotfix: Expect `redis_ns` key in connection config hash, not `ns`
This fixes a mistake that was added when allowing a `Hash` to be
passed directly to a `Connection` as its config. We expect a
`redis_ns` on a config object but the looked for `ns` in the hash.
This wasn't intended and they should be consistent. This fixes
the config building to look for the `redis_ns` key in the hash.

/cc @kellyredding