Update `ConnectionSpy`, handle `pipelined` and `multi`
This updates the `ConnectionSpy` to handle `pipelined` and `multi`
redis calls separately from all other redis calls. They will now
yield to the block passed to them which allows any redis calls
done in the block to automatically be spied as expected. This
makes it so the blocks do not have to manually called.

The `pipelined` and `multi` commands are special commands that
take a block in ruby. The `pipelined` command is for sending lots
of commands in a single request to the redis server. The benefit
is not paying the overhead of a round trip for each command. The
`multi` command is for running transactions in redis. The spy now
tracks the `pipelined` or `multi` call and then yields itself.
This makes it so a test can easily see that pipelining or a
transaction was used with a set of commands.