hotfix) backfill Array#sample for 1.8.7 compatibility
This backfills Array#sample so it works in ruby 1.8.7. This is
part of making sure the gem works for newer versions of ruby. This
allows the test suite to be run on newer versions of ruby which
can be used to tell if the gem is compatible with the version or
not.

Note: this also updates to the latest Assert which is fully
compatible with the newer ruby versions.