rework the overall API
This corrects some annoyances with HellaRedis, especially related
to directly using ::ConnectionPool's api.  These are breaking
changes that will require a major release.

Here is a summary of the changes:

* renamed Connection and ConnectionSpy to ConnetionPool and
  ConnectionPoolSpy.  These names more accurately describe what
  the object represent.  I also renamed RedisSpy to ConnectionSpy
  for similar reasons.
* Updated the ConnectionPool to be its own class that composes a
  ::ConnectionPool.  This allows me to define better method names
  for the api I want to expose and also not expose all of
  ::ConnectionPool's api.
* Switch to using a `connection` method instead of the crappy
  `with` method ::ConnectionPool provides.  This makes more
  intuitive sense about what object will be yielded to the given
  proc.
* Updated the main HellaRedis namespace to have a `new` method
  for building connection pool's from raw config args.  This
  allowed me to take out args/class handling from the pool classes
  and also allowed me to add a hook for test mode handling.
* added test mode handling to auto build pools vs spies depending
  on an environment variable.  This should allow for less manual
  spy stubbing in test code that uses HellaRedis.

Finally, I just reorganized everything to try and put each class
in its own file with its own tests and modernized the test suite
to get up to out latest conventions (factories, etc).  The one
exception is the Config is still under the HellaRedis namespace
and also in the  main hella_redis file.