add `reset!` method to connection pool spy
This allows clearing out all of the calls on a pool spy.  This is
handy to do in test setup/teardown so that call state from one
test doesn't interfere with other tests.

I chose to make the `calls` method on the connection spy an
accessor so I could manually empty it instead of making a `reset!`
method on the connection spy.  I didn't want to add another method
to the connection spy since it uses method missing to proxy the
redis instance.  I didn't want to pollute the method-space or add
possibilities for method collisions with the instance.