(hotfix) README cleanups
This updates the test mode section of the README to correct a
problem: the commands are always stored as symbols on the spies.
I noticed this when I was running some tests in an app and my
tests failed b/c I had expected a string (b/c the docs said to).

This also adds a bit to the test mode docs to illustrate how values
can be stubbed in to make sure you get a value you expect in your
tests.  I always forget what to do when I need a specific value
returned for a test and have to go lookup where I've previously
done it in the app.